move from rug to num-bigint (#53)

+++ src/pasta.rs
@@ -6,6 +6,8 @@ use crate::{
 use core::ops::Mul;
 use ff::Field;
 use merlin::Transcript;
+use num_bigint::BigInt;
+use num_traits::Num;
 use pasta_curves::{
   self,
   arithmetic::{CurveAffine, CurveExt, Group as Grp},
@@ -14,7 +16,6 @@ use pasta_curves::{
 };
 use rand::SeedableRng;
 use rand_chacha::ChaCha20Rng;
-use rug::Integer;
 
 //////////////////////////////////////Pallas///////////////////////////////////////////////
 
@@ -75,8 +76,8 @@ impl Group for pallas::Point {
     }
   }
 
-  fn get_order() -> Integer {
-    Integer::from_str_radix(
+  fn get_order() -> BigInt {
+    BigInt::from_str_radix(
       "40000000000000000000000000000000224698fc0994a8dd8c46eb2100000001",
       16,
     )
@@ -163,8 +164,8 @@ impl Group for vesta::Point {
     }
   }
 
-  fn get_order() -> Integer {
-    Integer::from_str_radix(
+  fn get_order() -> BigInt {
+    BigInt::from_str_radix(
       "40000000000000000000000000000000224698fc094cf91b992d30ed00000001",
       16,
     )
